// src/gen-schema/predefined-map.schema.d.ts
/* tslint:disable */
/**
 * This file was automatically generated by json-schema-to-typescript.
 * DO NOT MODIFY IT BY HAND. Instead, modify the source JSONSchema file,
 * and run json-schema-to-typescript to regenerate this file.
 */

export type PredefinedMapModel_Color = string;
export type PredefinedMapModel_TileType =
  | "FLOOR"
  | "FLOOR_HALL"
  | "NONE"
  | "STAIRS_DOWN"
  | "WALL"
  | "WALL_HALL"
  | "WALL_TOP";

export interface PredefinedMapModel {
  imageFilename: string;
  levelNumber: number;
  music?: string;
  startingPointColor: PredefinedMapModel_Color;
  tileColors: PredefinedMapModel_ColorMap;
  enemyColors: PredefinedMapModel_ColorMap;
  equipmentColors: PredefinedMapModel_ColorMap;
  itemColors: PredefinedMapModel_ColorMap;
  doorColors?: PredefinedMapModel_ColorMap;
  spawnerColors?: PredefinedMapModel_ColorMap;
  defaultTile?: PredefinedMapModel_TileType;
  tileset: string;
}
export interface PredefinedMapModel_ColorMap {
  /**
   * This interface was referenced by `PredefinedMapModel_ColorMap`'s JSON-Schema definition
   * via the `patternProperty` "^[A-Z_]+$".
   */
  [k: string]: string;
}